made it possible to copy module data into vec
