+++ src/internal/mod.rs
@@ -4,6 +4,8 @@ pub mod internal;
 pub mod utils;
 #[cfg(feature = "internal")]
 pub mod pattern_scan;
+#[cfg(feature = "internal")]
+pub mod module;
 
 #[cfg(feature = "internal")]
 pub use internal::*;
